+++ pg_data_etl/helpers/commands.py
@@ -1,0 +1,14 @@
+import subprocess
+
+
+def _run_command_in_shell(command: str) -> str:
+    """Use subprocess to execute a command in a shell"""
+
+    process = subprocess.Popen(command, shell=True, stdout=subprocess.PIPE)
+
+    output = process.communicate()
+    print(output[0])
+
+    process.kill()
+
+    return output
